requirements.txt
flask
google-generativeai
langchain
langchain-community
chromadb
pypdf
python-docx
sentence-transformers
faiss-cpu
tiktoken
python-dotenv
gunicorn